refactor: extract duplicated tab structure to eliminate code duplication
- Create shared TextGenerationTabContent component to eliminate 300+ lines of duplication
- Extract common tab structure (Prompt, Model, Input, Tools) into reusable component
- Maintain full functionality including model provider switching and tool management
- Fix TypeScript types with proper generic constraints
- Follow DRY principles and improve maintainability
- All type checking passes and builds successfully
