+++ .gitignore
@@ -43,3 +43,4 @@ yarn-error.log*
 
 # prepare fonts programmatically
 app/fonts/
+.env*.local
